Added authentication and authorisation

+++ backend/db/repository/blog.py
@@ -2,7 +2,7 @@ from sqlalchemy.orm import Session
 from schemas.blog import CreateBlog,UpdateBlog
 from db.models.blog import Blog
 
-def create_new_blog(blog:CreateBlog, db:Session, author_id:int =1):
+def create_new_blog(blog:CreateBlog, db:Session, author_id:int):
     blog = Blog(title=blog.title,
                 slug = blog.slug,
                 content = blog.content,
@@ -20,10 +20,12 @@ def list_blogs(db: Session):
     blogs = db.query(Blog).filter(Blog.is_active==True).all()
     return blogs
 
-def update_blog_by_id(id:int, blog: UpdateBlog, db:Session,author_id:int = 1):
+def update_blog_by_id(id:int, blog: UpdateBlog, db:Session,author_id:int):
     blog_in_db = db.query(Blog).filter(Blog.id==id).first()
     if not blog_in_db:
-        return
+        return {"error": f"Blog with id {id} does not exist"}
+    if not blog_in_db.author_id == author_id:
+        return {"error": f"Only the author can modify the blog"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content 
     db.add(blog_in_db)
@@ -32,6 +34,8 @@ def update_blog_by_id(id:int, blog: UpdateBlog, db:Session,author_id:int = 1):
 
 def delete_blog_by_id(id:int, db:Session, author_id: int):
     blog_in_db = db.query(Blog).filter(Blog.id==id).first()
+    if not blog_in_db.author_id == author_id:
+        return {"error": f"Only the author can modify the blog"}
     if not blog_in_db:
         return {"error": f"could not find a blog with id {id}"}
     db.delete(blog_in_db)
